Full rewrite (#14)
This release is a full rewrite of the `errors` package. See the README for further details.
#### Added
* `Caller` interface
* `Error` interface
* Exported methods
  - `Errorf(msg string, data ...interface{}) Error`
  - `GetCaller(err error) Caller`
  - `Has(err, test error) bool`
  - `Is(err, test error) bool`
  - `Trace(e error) Error`
  - `Track(e error) Error`

#### Changed
* Exported methods
  - `New(code std.Code, msg string, data ...interface{}) *Err`
    - `New(msg string) Error`
  - `Wrap(err error, code std.Code, msg string, data ...interface{}) *Err`
    - `Wrap(e error, msg string, data ...interface{}) Error`
#### Removed
* Exported methods
  - `From(code std.Code, err error) *Err`
* Support for error codes
* Support for sanitized vs raw error messages
* Support for HTTP status codes

+++ .travis.sh
@@ -1,9 +1,25 @@
 #!/bin/sh
 set -e
 
+exit_code=0
+
+go get -u golang.org/x/lint/golint
+[ "0" = "$?" ] || exit 1
+
+for dir in $(go list ./... | grep -v vendor); do
+    echo "golint $dir"
+    result=$(golint $dir)
+    if [ "" != "$result" ]; then
+        echo $result
+        exit_code=5
+    fi
+    if [ "0" != "$exit_code" ]; then
+        exit $exit_code
+    fi
+done
+
 rm -f coverage.txt
-for dir in $(go list ./...); do
-    echo "go test -timeout 20s -coverprofile=profile.out $dir"
+for dir in $(go list ./... | grep -v vendor); do
     go test -race -timeout 20s -coverprofile=profile.out $dir
     exit_code=$?
     if [ "0" != "$exit_code" ]; then
@@ -14,4 +30,5 @@ for dir in $(go list ./...); do
         rm profile.out
     fi
 done
+
 exit $exit_code
